split out individual components
