+++ src/pages/groups/GroupListPage.tsx
@@ -7,7 +7,11 @@ import { DelayedSearchInput } from 'src/components/input';
 import LeaveResourceModal from 'src/components/LeaveResourceModal';
 import { PageBox, PageBoxVariants } from 'src/components/PageBox';
 import TopBar from 'src/components/TopBar';
-import { GroupCard, GroupCardHeaders, NewGroupCard, NewGroupModal, NoGroupsMessage } from 'src/groups/List';
+import { GroupCard } from 'src/groups/GroupCard';
+import { GroupCardHeaders } from 'src/groups/GroupCardHeaders';
+import { NewGroupCard } from 'src/groups/NewGroupCard';
+import { NewGroupModal } from 'src/groups/NewGroupModal';
+import { NoGroupsMessage } from 'src/groups/NoGroupsMessage';
 import { Ajax } from 'src/libs/ajax';
 import { CurrentUserGroupMembership } from 'src/libs/ajax/Groups';
 import { withErrorReporting } from 'src/libs/error';
